Rudimentary support for json format juniper get_config
If no format attribute is specified get_config returns config as a plain text, but in json is specified, it returns a dict. 
Probably it still needs further polishing. e.g I cast it to a str, but maybe i could stay as dict.  It's my first pull request ever, Sorry if I messed something up.